Updating the template layouts to make it easier for refactoring later
